login now returns jwt token,dto constructor fixed

+++ fishing-booker/src/main/java/com/ftn/fishingbooker/controller/AuthenticationController.java
@@ -71,8 +71,7 @@ public class AuthenticationController {
 
             User user = (User) authentication.getPrincipal();
             String jwt = tokenUtils.generateToken(user.getEmail(), user.getRole().getName(), false);
-            int expiresIn = tokenUtils.getExpiredIn();
-
+            Long expiresIn = tokenUtils.getExpiredIn();
             return ResponseEntity.ok(new UserTokenStateDto(jwt, expiresIn));
         } catch (BadCredentialsException e) {
             throw new ResourceConflictException("Bad Credentials!");
